Title: 사각형 채우기 3, Time: 7ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 250119/사각형 채우기 3/rectangle-fill-3.cpp
@@ -8,12 +8,12 @@ int n;
 // dp[n] = dp[n-1] * 2 + dp[n - 2] * 3; 
 int main() {
     cin >> n;
-    int dp[1001] = { 0 } ; 
+    long long dp[1001] = { 0 } ; 
     dp[0] = 1;
     dp[1] = 2; 
     dp[2] = 7; 
 
-    const int DIV = 1000000007;
+    const long long DIV = 1000000007;
     for(int i = 3; i <= n; i ++) { 
         dp[i] = (dp[i-1] * 2 % DIV+ dp[i-2] * 3 % DIV) % DIV;
         if(i % 2 == 1) {
